+++ EasyAccess/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Easy Access Change Log
 
+## 2.1.5 (Unreleased)
+
+### Changed
+
+* If config file is missing, it will attempt to restore from global data.
+
 ## 2.1.4 (April 12, 2024)
 
 ### Changed