+++ ExpandedStorage/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Expanded Storage Change Log
 
+## 3.0.6 (Unreleased)
+
+### Changed
+
+* If config file is missing, it will attempt to restore from global data.
+
 ## 3.0.5 (April 12, 2024)
 
 ### Changed